Extract logic to change manga in library state into hook
